+++ CITATION.cff
@@ -1,5 +1,5 @@
 cff-version: 1.2.0
-message: "If you use PyHd, please cite it as below."
+message: "If you use PyHeatDemand, please cite it as below."
 authors:
   - family-names: "Jüstel"
     given-names: "Alexander"
@@ -8,7 +8,7 @@ authors:
     given-names: "Frank"
     orcid: ames: "Alexander"
     orcid: https://orcid.org/0000-0002-3067-831X
-title: "PyHD - PyHeatDemand - Processing Tool for Heat Demand Data"
+title: "PyHeatDemand - Processing Tool for Heat Demand Data"
 version: 0.0.1
 doi: https://doi.org/10.21105/joss.xxxx
 date-released: 2023-10-31
