+++ docs/source/index.rst
@@ -1,22 +1,22 @@
-.. PyHD documentation master file, created by
+.. PyHeatDemand documentation master file, created by
    sphinx-quickstart on Fri Oct 20 20:40:17 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to documentation page of PyHD!
-======================================
+Welcome to documentation page of PyHeatDemand!
+==============================================
 
-**PyHD** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for
+**PyHeatDemand** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for
 constructing heat demand maps (raster). Knowledge about the heat demand (MWh/area/year) of a respective building,
 district, city, state, country or even on a continental scale is crucial for an adequate heat demand planning or
 planning for providing power plant capacities. Mapping of the heat demand may also identify potential areas for new
 district heating networks or even geothermal power plants for a climate-friendly heat production.
 
-This documentation contains information about the open-source package PyHD - PyHeatDemand - Processing Tool for Heat Demand Data.
+This documentation contains information about the open-source package PyHeatDemand - Processing Tool for Heat Demand Data.
 
 .. container:: button
 
-    :doc:`About PyHD <about>` | :doc:`Installation <installation>` | :doc:`API Reference <pyhd>` | :doc:`Calculating Heat Demand <heatdemand>` | `GitHub Repository <https://github.com/AlexanderJuestel/pyhd>`_ |
+    :doc:`About PyHeatDemand <about>` | :doc:`Installation <installation>` | :doc:`API Reference <pyheatdemand>` | :doc:`Calculating Heat Demand <heatdemand>` | `GitHub Repository <https://github.com/AlexanderJuestel/pyhd>`_ |
 
 
 .. toctree::
