+++ docs/source/modules.rst
@@ -4,4 +4,4 @@ API Reference
 .. toctree::
    :maxdepth: 4
 
-   pyhd
+   pyheatdemand
