--- docs/source/pyhd.rst
@@ -1,17 +1,0 @@
-.. _pyhd_ref:
-
-PyHD API Reference
-==================
-
-
-PyHD - PyHeatDemand - Processing
----------------------------------
-
-The API reference provides an overview of all functions and methods implemented in PyHD.
-
-.. automodule:: pyhd.processing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
