+++ docs/source/pyheatdemand.rst
@@ -1,0 +1,17 @@
+.. _pyhd_ref:
+
+PyHeatDemand API Reference
+===========================
+
+
+PyHeatDemand - Processing
+---------------------------------
+
+The API reference provides an overview of all functions and methods implemented in PyHeatDemand.
+
+.. automodule:: pyheatdemand.processing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
