+++ environment_dev.yml
@@ -1,4 +1,4 @@
-name: pyhd
+name: pyheatdemand
 channels:
   - conda-forge
 dependencies:
